Change data type double => decimal

+++ Nexttech/Models/Calculation.cs
@@ -8,51 +8,51 @@ using System.ComponentModel.DataAnnotations;
     public int MaterialId { get; set; }
     public int PartsProduced { get; set; }
     public int NumberOfBuilds { get; set; }
-    public double PartMass { get; set; }
-    public double PartHeight { get; set; }
-    public double PartArea { get; set; }
-    public double SupportMat { get; set; }
+    public decimal PartMass { get; set; }
+    public decimal PartHeight { get; set; }
+    public decimal PartArea { get; set; }
+    public decimal SupportMat { get; set; }
 
-    public double TotalMaterial { get; set; }
-    public double SupportMass { get; set; }
-    public double TotalSupport { get; set; }
-    public double TotalMaterialAllBuilds { get; set; }
-    public double Recycled { get; set; }
-    public double Waste { get; set; }
-    public double RequiredMaterial { get; set; }
-    public double MaterialCost { get; set; }
+    public decimal TotalMaterial { get; set; }
+    public decimal SupportMass { get; set; }
+    public decimal TotalSupport { get; set; }
+    public decimal TotalMaterialAllBuilds { get; set; }
+    public decimal Recycled { get; set; }
+    public decimal Waste { get; set; }
+    public decimal RequiredMaterial { get; set; }
+    public decimal MaterialCost { get; set; }
 
-    public double PrepTime { get; set; }
-    public double BuildPrepCost { get; set; }
+    public decimal PrepTime { get; set; }
+    public decimal BuildPrepCost { get; set; }
 
-    public double PostTimePerPart { get; set; }
-    public double TotalPostTime { get; set; }
-    public double PostProcessCost { get; set; }
+    public decimal PostTimePerPart { get; set; }
+    public decimal TotalPostTime { get; set; }
+    public decimal PostProcessCost { get; set; }
 
-    public double UpFront { get; set; }
-    public double AnnualDepreciation { get; set; }
-    public double AnnualMaintenance { get; set; }
-    public double AnnualMachineCost { get; set; }
-    public double HoursPerYear { get; set; }
-    public double MachineCostPerHour { get; set; }
+    public decimal UpFront { get; set; }
+    public decimal AnnualDepreciation { get; set; }
+    public decimal AnnualMaintenance { get; set; }
+    public decimal AnnualMachineCost { get; set; }
+    public decimal HoursPerYear { get; set; }
+    public decimal MachineCostPerHour { get; set; }
 
-    public double WarmupTotal { get; set; }
-    public double PartVolume { get; set; }
-    public double PrintTime { get; set; }
-    public double CooldownTotal { get; set; }
-    public double ExchangeTime { get; set; }
-    public double MachineTime { get; set; }
-    public double MachineCost { get; set; }
+    public decimal WarmupTotal { get; set; }
+    public decimal PartVolume { get; set; }
+    public decimal PrintTime { get; set; }
+    public decimal CooldownTotal { get; set; }
+    public decimal ExchangeTime { get; set; }
+    public decimal MachineTime { get; set; }
+    public decimal MachineCost { get; set; }
 
-    public double OperatingTotalCost { get; set; }
-    public double Consumables { get; set; }
-    public double ConsumablesCost { get; set; }
+    public decimal OperatingTotalCost { get; set; }
+    public decimal Consumables { get; set; }
+    public decimal ConsumablesCost { get; set; }
 
-    public double LaborBuildTime { get; set; }
-    public double LaborExchangeTime { get; set; }
-    public double LaborCost { get; set; }
+    public decimal LaborBuildTime { get; set; }
+    public decimal LaborExchangeTime { get; set; }
+    public decimal LaborCost { get; set; }
 
-    public double TotalCost { get; set; }
+    public decimal TotalCost { get; set; }
 
     public DateTime CreatedAt { get; set; }
